+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Miscellaneous:
 - Moved randomly generated test vectors to individual subfolders.
 - Randomly generated test vectors are no longer a part of the default test suite, because they cannot be tested consistently, because they are dependent on specific versions of tzdata.
 - Use independently configured directories for tzdata and vectors in the test runner.
+- Display configuration parameters and the list of loaded test vector files in the test runner.
 
 ## 2.0 (2016-07-19)